Refactor render frame execution into render_manifest
Move begin/end_render_frame, execute_manifest, barrier
emission, and helper functions from renderer.cpp into
render_manifest.cpp. Expose handle_window_resize and
get_or_create_framebuffer as non-intern functions.
Move UBO structs to render_manifest.h. Rename
static_model to static_mesh. Fix asset_flags u64->u32.

+++ src/asset_defs.h
@@ -26,7 +26,7 @@ enum asset_flag : u32
     ASSET_FLAG_LOADED = make_flag(1),
     ASSET_FLAG_USER_BASE = make_flag(2),
 };
-using asset_flags = u64;
+using asset_flags = u32;
 
 
 const sizet ASSET_POOL_MEMORY_BUDGETS[ASSET_TYPE_USER] = {
